+++ DataRepo/utils/__init__.py
@@ -1,17 +1,6 @@
 from DataRepo.utils.accucor_data_loader import (
     AccuCorDataLoader,
-    InvalidLCMSHeaders,
     IsotopeObservationData,
-    IsotopeObservationParsingError,
-    LCMethodFixturesMissing,
-    LCMSDefaultsRequired,
-    MismatchedSampleHeaderMZXML,
-    MissingLCMSSampleDataHeaders,
-    MissingMZXMLFiles,
-    NoMZXMLFiles,
-    PeakAnnotFileMismatches,
-    TracerLabeledElementNotFound,
-    UnexpectedLCMSSampleDataHeaders,
     lcms_headers_are_valid,
 )
 from DataRepo.utils.composite_data import leaderboard_data
@@ -26,17 +15,31 @@ from DataRepo.utils.exceptions import (
     DryRun,
     DupeCompoundIsotopeCombos,
     DuplicatePeakGroups,
+    DuplicateSampleDataHeaders,
     DuplicateValues,
     HeaderConfigError,
     HeaderError,
+    InvalidLCMSHeaders,
+    IsotopeObservationParsingError,
+    LCMethodFixturesMissing,
+    LCMSDefaultsRequired,
     LoadingError,
+    MismatchedSampleHeaderMZXML,
     MissingCompounds,
+    MissingLCMSSampleDataHeaders,
+    MissingMZXMLFiles,
+    MissingPeakAnnotationFiles,
+    MissingRequiredLCMSValues,
     MissingSamplesError,
     MissingTissues,
+    NoMZXMLFiles,
     NoSamplesError,
+    PeakAnnotFileMismatches,
     RequiredSampleValuesError,
     RequiredValueError,
     SheetMergeError,
+    TracerLabeledElementNotFound,
+    UnexpectedLCMSSampleDataHeaders,
     UnknownHeadersError,
     UnskippedBlanksError,
 )
@@ -108,4 +111,7 @@ __all__ = [
     "parse_infusate_name",
     "parse_tracer_concentrations",
     "ProtocolsLoader",
+    "DuplicateSampleDataHeaders",
+    "MissingPeakAnnotationFiles",
+    "MissingRequiredLCMSValues",
 ]
